+++ server/src/enums/roles.enum.ts
@@ -1,7 +1,5 @@
 export const Roles = {
     SUPER: 'super',
 
-    GUEST: 'guest',
-
     MEMBER: 'member',
 };