+++ server/src/modules/auth/auth.service.ts
@@ -2,6 +2,7 @@ import Redis from 'ioredis';
 import ms from 'ms';
 import { ITokenPayload } from 'src/decorators/token-payload.decorators';
 import { LoginType } from 'src/enums/login-type.enum';
+import { Roles } from 'src/enums/roles.enum';
 import { verifyPasswordLogin } from 'src/utils/credential';
 import { getPermissionsByRoles } from 'src/utils/permissions';
 import { getBase64Uuid } from 'src/utils/uuid';
@@ -18,6 +19,7 @@ import { JwtService } from '@nestjs/jwt';
 import { CacheService } from '../../middlewares/cache.service';
 import { AccountAuth } from '../accounts/entities/account-auth.entity';
 import { Account } from '../accounts/entities/account.entity';
+import { Role } from '../roles/entities/role.entity';
 import { LoginOrCreateDto } from './dto/login-or-create.dto';
 import { OAuthService } from './oauth.service';
 
@@ -82,6 +84,8 @@ export class AuthService {
 
             // 找不到 auth data, 自動建立帳號
             if (!findAuth) {
+                const findMemberRole = await this.entityManager.findOne(Role, { where: { role: Roles.MEMBER } });
+
                 const newAccount = new Account({
                     email: oauthResult.email,
                     name: oauthResult.name,
@@ -92,6 +96,7 @@ export class AuthService {
                             credential: oauthResult.credential,
                         }),
                     ],
+                    roles: [findMemberRole],
                 });
 
                 id = (await this.entityManager.save(newAccount)).id;