+++ server/src/modules/auth/dto/login-or-create.dto.ts
@@ -2,6 +2,7 @@ import {
     IsEmail,
     IsNotEmpty,
     IsString,
+    ValidateIf,
 } from 'class-validator';
 import { LoginType } from 'src/enums/login-type.enum';
 
@@ -19,16 +20,23 @@ export class LoginOrCreateDto {
     })
         type: typeof LoginType[keyof typeof LoginType];
 
+    @ValidateIf((o) => o.type === LoginType.PASSWORD)
     @IsEmail()
     @IsNotEmpty()
-    @ApiProperty()
-        email: string;
+    @ApiProperty({
+        required: false, description: 'required when using password type',
+    })
+        email?: string;
 
+    @ValidateIf((o) => o.type === LoginType.PASSWORD)
+    @IsNotEmpty()
     @ApiProperty({
         required: false, description: 'required when using password type', 
     })
-        password: string;
+        password?: string;
 
+    @ValidateIf((o) => o.type === LoginType.GOOGLE)
+    @IsNotEmpty()
     @ApiProperty({
         required: false, description: 'required when not using password login type', 
     })